#38 extract wrapper class
